Revised some commentary.
I had an idea to try and merge the lexer and the parser.  It is, after
all, a PEG-generated lexer, and PEGs are supposed to generate parsers,
right?  But the elegance of the separation at this point is so nice and
straightforward that I don't think I win anything by merging them.